update Godeps of opencontainers/specs to 0.3.0
