+++ Godeps/_workspace/src/github.com/opencontainers/specs/.travis.yml
@@ -1,6 +1,6 @@
 language: go
 go:
-  - 1.5.1
+  - 1.5.3
   - 1.4.3
   - 1.3.3
